feat: Add GroupCard and HighLight components
This commit adds the GroupCard and HighLight components to the project. The GroupCard component is a reusable card component for displaying group information, while the HighLight component is used to highlight important information on the screen. These components will be used in the Groups screen to display group details and highlight important messages.