feat: Add ListEmpty component and update Players screen
The ListEmpty component is added to handle the case when there are no players in the team. The Players screen is updated to use the ListEmpty component and also includes a button to remove the team. This enhances the user experience by providing a clear message when there are no players and allowing the user to easily remove the team.